requirements.txt
async-timeout==4.0.3
bleak==0.22.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.6
dotmap==1.3.30
idna==3.7
meshtastic==2.3.8
packaging==24.0
paho-mqtt==2.1.0
pexpect==4.9.0
protobuf==5.26.1
ptyprocess==0.7.0
pycparser==2.22
pycryptodome==3.20.0
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-CoreBluetooth==10.2
pyobjc-framework-libdispatch==10.2
Pypubsub==4.0.3
PyQRCode==1.2.1
pyserial==3.5
PyYAML==6.0.1
requests==2.31.0
tabulate==0.9.0
timeago==1.0.16
tk==0.1.0
typing_extensions==4.11.0
urllib3==2.2.1
